Fix some jextract test failures

+++ test/jdk/tools/jextract/test8244412/test8244412.h
@@ -21,5 +21,5 @@
  * questions.
  */
 
-typedef long mysize_t;
-typedef long MYSIZE_T;
+typedef long long mysize_t;
+typedef long long MYSIZE_T;
